Refactor admin logic into AdminCommand superclass
