+++ app/models/hackbot/interactions/set_poc.rb
@@ -1,13 +1,11 @@
 module Hackbot
   module Interactions
-    class SetPoc < Command
+    class SetPoc < AdminCommand
       TRIGGER = /set-poc ?(?<streak_key>.+)/
 
       USAGE = 'set-poc <leader_streak_key>'.freeze
       DESCRIPTION = 'set the given leader as the point of contact for their '\
                     'club (staff only)'.freeze
-
-      before_handle :ensure_admin
 
       def start
         streak_key = captured[:streak_key]
@@ -129,14 +127,6 @@ module Hackbot
 
       def get_last_arg(text)
         text.split(' ').last
-      end
-
-      def ensure_admin
-        return if current_admin?
-
-        msg_channel(copy('not_admin'))
-
-        throw :abort
       end
     end
   end
