+++ lib/data/copy/admin_command.yml
@@ -1,0 +1,2 @@
+when_in_start_state: You must be an administrator to use this command.
+when_in_progress: You must be an administrator to interact with this command.
