+++ lib/data/copy/lookup.yml
@@ -1,4 +1,3 @@
-not_admin: You must be an admin to use this command.
 not_found: Unable to find a leader with that username.
 
 confirm:
